Move common params to abstract class
