+++ src/netspresso_trainer/pipelines/builder.py
@@ -88,6 +88,8 @@ def build_pipeline(pipeline_type, conf, task, model_name, model, devices,
                                             task_processor=task_processor,
                                             model_name=model_name,
                                             model=model,
+                                            logger=train_logger,
+                                            timer=timer,
                                             optimizer=optimizer,
                                             scheduler=scheduler,
                                             loss_factory=loss_factory,
@@ -96,8 +98,6 @@ def build_pipeline(pipeline_type, conf, task, model_name, model, devices,
                                             eval_dataloader=eval_dataloader,
                                             single_gpu_or_rank_zero=single_gpu_or_rank_zero,
                                             is_graphmodule_training=is_graphmodule_training,
-                                            logger=train_logger,
-                                            timer=timer,
                                             model_ema=model_ema,
                                             start_epoch=start_epoch,
                                             cur_epoch=cur_epoch,
